+++ frontend/src/components/Search/Results.module.css
@@ -7,27 +7,10 @@
   width: 100%;
 }
 
-.results_header_row {
-  user-select: none;
-}
-
-.one_line {
-  display: block;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
 .up_btn:hover {
   cursor: pointer;
   color: rgb(180, 180, 180);
   transform: scale(1.1);
-}
-
-.list_grid_toggle {
-  position: absolute;
-  left: 5px;
-  z-index: 2;
 }
 
 .icon {
