+++ frontend/src/components/Search/ResultsHeaders.module.css
@@ -1,0 +1,16 @@
+.results_header_row {
+  user-select: none;
+}
+
+.list_grid_toggle {
+  position: absolute;
+  left: 5px;
+  z-index: 2;
+}
+
+.one_line {
+  display: block;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
